chore: Move backend unit tests to new repo
- move `quizUnit.test.js` and `userUnit.test.js` to reside under `test/unit/backend/controller`, updating import paths accordingly for better organization and clarity.
